actions: Split out zerver.actions.user_topics.
